removing opt varibles
